It highlights matches
Now, it doesn't update the screen at the right times. I have to push submit
and then type something at the end of the Parse This. these are problems.
And it doesn't remove the highlighting when I change the Parse This.
that I could manage in state. but how do I get it to notice the change
once I get the tree? ... probably more state management, bummer